Conflicts:
	crypto/asn1/a_strnid.c
	crypto/objects/obj_dat.h
	crypto/objects/obj_mac.num
	crypto/objects/objects.txt
	include/openssl/obj_mac.h

+++ ssl/pqueue.c
@@ -141,10 +141,10 @@ pitem *pqueue_next(pitem **item)
     return ret;
 }
 
-int pqueue_size(pqueue *pq)
+size_t pqueue_size(pqueue *pq)
 {
     pitem *item = pq->items;
-    int count = 0;
+    size_t count = 0;
 
     while (item != NULL) {
         count++;
